Made class from separate functions
